Module Tezos_client_015_PtLimaPt.Client_proto_context

val list_contract_labels : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> (string * string * string) list Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_alpha.Protocol.Contract_services.list.

val get_storage : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> unparsing_mode: Tezos_protocol_015_PtLimaPt.Protocol.Script_ir_unparser.unparsing_mode -> Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.expr option Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_plugin_alpha.Plugin.RPC.Contract.get_storage_normalized.

val get_used_storage_space : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t -> Z.t option Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_plugin_alpha.Plugin.RPC.Contract.get_used_storage_space.

val get_paid_storage_space : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t -> Z.t option Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_plugin_alpha.Plugin.RPC.Contract.get_paid_storage_space.

val get_contract_big_map_value : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t -> (Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.expr * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.expr) -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.expr option Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_alpha.Protocol.Contract_services.contract_big_map_get_opt.

val register_global_constant : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> constant:string -> unit -> (Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.register_global_constant Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Injection.result, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

Calls Injection.prepare_manager_operation with Register_global_constant constant as operation.

val get_big_map_value : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> unparsing_mode: Tezos_protocol_015_PtLimaPt.Protocol.Script_ir_unparser.unparsing_mode -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Big_map.Id.t -> Tezos_protocol_015_PtLimaPt.Protocol.Script_expr_hash.t -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.expr Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_plugin_alpha.Plugin.RPC.Big_map.big_map_get_normalized.

val get_script : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> unparsing_mode: Tezos_protocol_015_PtLimaPt.Protocol.Script_ir_unparser.unparsing_mode -> normalize_types:bool -> Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.t option Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_plugin_alpha.Plugin.RPC.Contract.get_script_normalized.

val get_script_hash : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t -> Tezos_protocol_015_PtLimaPt.Protocol.Script_expr_hash.t option Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_alpha.Protocol.Contract_services.script_opt.

val get_balance : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Contract.t -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_alpha.Protocol.Contract_services.balance.

val get_frozen_deposits_limit : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> Tezos_base.TzPervasives.Signature.V0.Public_key_hash.t -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t option Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_alpha.Protocol.Delegate_services.frozen_deposits_limit.

val build_delegate_operation : ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash option -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.delegation Annotated_manager_operation.t

Calls Injection.prepare_manager_operation with Alpha_context.Delegation delegate_opt as operation.

val set_delegate : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> manager_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash option -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.delegation Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Injection.result Tezos_base.TzPervasives.tzresult Lwt.t

Calls Injection.inject_manager_operation with Annotated_manager_operation.annotated_list.Single_manager build_delegate_operation ?fee opt_delegate as operation.

val update_consensus_key : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> consensus_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> manager_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> Tezos_base.TzPervasives.Signature.V0.public_key -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.update_consensus_key Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Injection.result Tezos_base.TzPervasives.tzresult Lwt.t

val drain_delegate : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> consensus_sk:Tezos_client_base.Client_keys_v0.sk_uri -> consensus_pkh:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> ?destination:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> delegate:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> unit -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.drain_delegate Injection.result Tezos_base.TzPervasives.tzresult Lwt.t

val set_deposits_limit : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> manager_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t option -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.set_deposits_limit Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Injection.result Tezos_base.TzPervasives.tzresult Lwt.t

Calls Injection.inject_manager_operation with Annotated_manager_operation.annotated_list.Single_manager Alpha_context.Set_deposits_limit limit_opt as operation.

val increase_paid_storage : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?force:bool -> ?dry_run:bool -> ?verbose_signing:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?confirmations:int -> ?simulation:bool -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> destination:Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> manager_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> amount_in_bytes:Z.t -> unit -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.increase_paid_storage Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Injection.result Tezos_base.TzPervasives.tzresult Lwt.t

Calls Injection.inject_manager_operation with Annotated_manager_operation.annotated_list.Single_manager Alpha_context.Increase_paid_storage {amount_in_bytes; destination} as operation.

val register_as_delegate : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> manager_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> ?consensus_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> (Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.delegation Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Injection.result * (Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.update_consensus_key Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.update_consensus_key Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) option) Tezos_base.TzPervasives.tzresult Lwt.t

Same as set_delegate but the ~source argument of Injection.inject_manager_operation is Signature.V0.Public_key.hash src_pk.

val save_contract : force:bool -> Protocol_client_context.full -> string -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Contract.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Calls Raw_contract_alias.add.

val originate_contract : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?branch:int -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> delegate: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash option -> initial_storage:string -> balance:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> code:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.expr -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.origination Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Injection.result * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Contract.t) Tezos_base.TzPervasives.tzresult Lwt.t

Injects the origination of a script into the context. See Injection.inject_manager_operation for the injection, and Protocol.Alpha_context.Origination for the origination parameters.

val parse_arg_transfer : string option -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.lazy_expr Tezos_base.TzPervasives.tzresult Lwt.t

Calls Michelson_v1_parser.parse_expressionarg.

val build_transaction_operation : amount:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> parameters: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.lazy_expr -> ?entrypoint:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Entrypoint.t -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Contract.t -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.transaction Annotated_manager_operation.t

Calls Injection.prepare_manager_operation with Alpha_context.Transaction {amount;parameters;destinations;entrypoint} as operation.

val transfer_with_script : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?force:bool -> ?branch:int -> ?successor_level:bool -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> destination:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Contract.t -> ?entrypoint:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Entrypoint.t -> parameters: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.lazy_expr -> amount:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> fee_parameter:Injection.fee_parameter -> ?replace_by_fees:bool -> unit -> (Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.transaction Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Injection.result * Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t list) Tezos_base.TzPervasives.tzresult Lwt.t

Same as transfer, but takes parameters as Script.lazy_expr instead of a raw string. This can be useful for commands that elaborate Micheline parameters, such as Multisigs or FA1.2 operations.

val transfer : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?force:bool -> ?branch:int -> ?successor_level:bool -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> destination:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Contract.t -> ?entrypoint:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Entrypoint.t -> ?arg:string -> amount:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> fee_parameter:Injection.fee_parameter -> ?replace_by_fees:bool -> unit -> (Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.transaction Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Injection.result * Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t list) Tezos_base.TzPervasives.tzresult Lwt.t

Calls Injection.inject_manager_operation with Annotated_manager_operation.annotated_list.Single_manager build_transaction_operation ~amount ~parameters ~entrypoint ?fee ?gas_limit ?storage_limit destination as contents.

val build_reveal_operation : ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.reveal Annotated_manager_operation.t

Calls Injection.prepare_manager_operation with Reveal pk as operation

val reveal : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?branch:int -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> fee_parameter:Injection.fee_parameter -> unit -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.reveal Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Injection.result Tezos_base.TzPervasives.tzresult Lwt.t

Calls Injection.inject_manager_operation with Annotated_manager_operation.annotated_list.Single_manager build_reveal_operation ?fee ~storage_limit:Z.zero src_pk as contents.

type activation_key = {

pkh : Tezos_base.TzPervasives.Signature.Ed25519.Public_key_hash.t;

amount : Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t;

activation_code : Tezos_protocol_015_PtLimaPt.Protocol.Blinded_public_key_hash.activation_code;

mnemonic : string list;

password : string;

email : string;

}

val activation_key_encoding : activation_key Tezos_base.TzPervasives.Data_encoding.t

type batch_transfer_operation = {

destination : string;

fee : string option;

gas_limit : Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral option;

storage_limit : Z.t option;

amount : string;

arg : string option;

entrypoint : Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Entrypoint.t option;

}

val batch_transfer_operation_encoding : batch_transfer_operation Tezos_base.TzPervasives.Data_encoding.t

val activate_account : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?encrypted:bool -> ?force:bool -> activation_key -> string -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.activate_account Injection.result Tezos_base.TzPervasives.tzresult Lwt.t

Activate an account, by calling Injection.inject_operation.

val activate_existing_account : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> string -> Tezos_protocol_015_PtLimaPt.Protocol.Blinded_public_key_hash.activation_code -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.activate_account Injection.result Tezos_base.TzPervasives.tzresult Lwt.t

Activate an existing account, by calling Injection.inject_operation with activation code. It fails if the account is unknown or if the account is not Ed25519.

type period_info = {

current_period_kind : Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Voting_period.kind;

position : Stdlib.Int32.t;

remaining : Stdlib.Int32.t;

current_proposal : Tezos_base.TzPervasives.Protocol_hash.t option;

}

type ballots_info = {

current_quorum : Stdlib.Int32.t;

participation : Stdlib.Int32.t;

supermajority : Stdlib.Int64.t;

ballots : Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Vote.ballots;

}

val get_period_info : ?successor:bool -> Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> period_info Tezos_base.TzPervasives.tzresult Lwt.t

get_period_info ~successor cctx ~chain ~block returns the successor period_info if successor, it returns the current period_info otherwise. This function uses Tezos_protocol_alpha.Protocol.Voting_services.successor_period if successor, otherwise it calls Tezos_protocol_alpha.Protocol.Voting_services.current_period. In any case, it also uses Tezos_protocol_alpha.Protocol.Voting_services.current_proposal

val get_ballots_info : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ballots_info Tezos_base.TzPervasives.tzresult Lwt.t

get_ballots_info cctx ~chain ~block returns the ballots_info. It calls Tezos_protocol_alpha.Protocol.Voting_services.ballots, Tezos_protocol_alpha.Protocol.Voting_services.current_quorum and Tezos_protocol_alpha.Protocol.Voting_services.total_voting_power.

val get_proposals : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> Stdlib.Int64.t Tezos_protocol_015_PtLimaPt.Environment.Protocol_hash.Map.t Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_alpha.Protocol.Voting_services.proposals

val submit_proposals : ?dry_run:bool -> ?verbose_signing:bool -> Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> Tezos_base.TzPervasives.Protocol_hash.t list -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.proposals Injection.result_list Tezos_base.TzPervasives.tzresult Lwt.t

Calls Injection.inject_operation where contents is Alpha_context.Single (Proposals {source; period = index; proposals}) with index the result of Alpha_services.Voting.successor_period

val submit_ballot : ?dry_run:bool -> ?verbose_signing:bool -> Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> Tezos_base.TzPervasives.Protocol_hash.t -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Vote.ballot -> Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.ballot Injection.result_list Tezos_base.TzPervasives.tzresult Lwt.t

Calls Injection.inject_operation where contents is Alpha_context.Single (Ballot {source; period = index; proposals}) with index the result of Alpha_services.Voting.successor_period

val display_receipt_for_operation : Protocol_client_context.full -> chain:Tezos_shell_services.Block_services.chain -> ?predecessors:int -> Tezos_base.TzPervasives.Operation_list_hash.elt -> unit Tezos_base.TzPervasives.tzresult Lwt.t

lookup an operation in predecessors previous blocks, and print the receipt if found

val cached_contracts : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> (Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t * int) list Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_alpha.Protocol.Alpha_services.Cache.cached_contracts

val contract_rank : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> Tezos_protocol_015_PtLimaPt.Protocol.Contract_hash.t -> int option Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_alpha.Protocol.Alpha_services.Cache.contract_rank

val contract_cache_size : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> int Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_alpha.Protocol.Alpha_services.Cache.contract_cache_size

val contract_cache_size_limit : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> int Tezos_base.TzPervasives.tzresult Lwt.t

Calls Tezos_protocol_alpha.Protocol.Alpha_services.Cache.contract_cache_size_limit

val originate_tx_rollup : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_origination Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_origination Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val submit_tx_rollup_batch : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?burn_limit:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> content:string -> tx_rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup.t -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_submit_batch Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_submit_batch Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val submit_tx_rollup_commitment : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> level:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_level.t -> inbox_merkle_root: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_inbox.Merkle.root -> messages: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_message_result_hash.t list -> predecessor: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_commitment_hash.t option -> tx_rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup.t -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_commit Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_commit Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val submit_tx_rollup_finalize_commitment : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> tx_rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup.t -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_finalize_commitment Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_finalize_commitment Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val submit_tx_rollup_remove_commitment : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> tx_rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup.t -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_remove_commitment Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_remove_commitment Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val submit_tx_rollup_rejection : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> level:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_level.t -> tx_rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup.t -> message: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_message.t -> message_position:int -> message_path: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_inbox.Merkle.path -> message_result_hash: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_message_result_hash.t -> message_result_path: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_commitment.Merkle.path -> previous_context_hash:Tezos_base.TzPervasives.Context_hash.t -> previous_withdraw_list_hash: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_withdraw_list_hash.t -> previous_message_result_path: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_commitment.Merkle.path -> proof:Tezos_protocol_015_PtLimaPt.Protocol.Tx_rollup_l2_proof.t -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_rejection Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_rejection Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val submit_tx_rollup_return_bond : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> tx_rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup.t -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_return_bond Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_return_bond Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val tx_rollup_dispatch_tickets : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> level:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_level.t -> context_hash:Tezos_base.TzPervasives.Context_hash.t -> message_position:int -> message_result_path: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_commitment.Merkle.path -> tickets_info: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup_reveal.t list -> tx_rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tx_rollup.t -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_dispatch_tickets Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.tx_rollup_dispatch_tickets Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val transfer_ticket : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> contents:string -> ty:string -> ticketer:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Contract.t -> amount:Tezos_protocol_015_PtLimaPt.Protocol.Ticket_amount.t -> destination:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Contract.t -> entrypoint:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Entrypoint.t -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.transfer_ticket Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.transfer_ticket Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val sc_rollup_originate : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> ?counter:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> kind:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.Kind.t -> boot_sector:string -> parameters_ty: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Script.lazy_expr -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_originate Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_originate Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val sc_rollup_add_messages : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> ?counter:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.t -> messages:string list -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_add_messages Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_add_messages Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val sc_rollup_cement : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> ?counter:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.t -> commitment: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.Commitment.Hash.t -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_cement Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_cement Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val sc_rollup_publish : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> ?counter:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.t -> commitment: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.Commitment.t -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_publish Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_publish Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val sc_rollup_execute_outbox_message : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> ?counter:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.t -> cemented_commitment: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.Commitment.Hash.t -> output_proof:string -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_execute_outbox_message Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_execute_outbox_message Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val sc_rollup_recover_bond : Protocol_client_context.full -> chain:Tezos_shell_services.Shell_services.chain -> block:Tezos_shell_services.Shell_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.tez -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?counter:Z.t -> source:Tezos_base.TzPervasives.Signature.V0.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.V0.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> sc_rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.t -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_recover_bond Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_recover_bond Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val sc_rollup_refute : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> ?counter:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.t -> refutation: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.Game.refutation option -> opponent: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.Staker.t -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_refute Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_refute Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val sc_rollup_timeout : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> ?counter:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.t -> alice:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.Staker.t -> bob:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.Staker.t -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_timeout Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_timeout Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t

val sc_rollup_dal_slot_subscribe : Protocol_client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> ?confirmations:int -> ?dry_run:bool -> ?verbose_signing:bool -> ?simulation:bool -> ?fee:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Tez.t -> ?gas_limit: Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> ?counter:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.counter -> source:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key_hash -> rollup:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Sc_rollup.t -> slot_index:int -> src_pk:Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.public_key -> src_sk:Tezos_client_base.Client_keys_v0.sk_uri -> fee_parameter:Injection.fee_parameter -> unit -> (Tezos_base.TzPervasives.Operation_hash.t * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_dal_slot_subscribe Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.contents * Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.sc_rollup_dal_slot_subscribe Tezos_protocol_015_PtLimaPt.Protocol.Alpha_context.Kind.manager Tezos_protocol_015_PtLimaPt.Protocol.Apply_results.contents_result) Tezos_base.TzPervasives.tzresult Lwt.t